refactor(ST-46): renamed components and fixed builds by removing unnecessary components
