--- modules/classroom/components/people/addStudentOnClassroom/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddStudentOnClassroom";
