--- modules/classroom/components/people/addStudentOnClassroomForm/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddStudentOnClassForm";
