+++ modules/classroom/components/people/addStudentToClassroom/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AddStudentToClassroom";
