+++ modules/classroom/containers/classroomContainer/ClassroomContainer.tsx
@@ -5,7 +5,6 @@ import { useRouter } from "next/router";
 import { Tabs } from "@mantine/core";
 
 import PeopleContainer from "@/modules/classroom/containers/peopleContainer";
-import StreamContainer from "@/modules/classroom/containers/streamContainer";
 import LoadingComponent from "@/shared/components/LoadingComponent";
 import { useLazyGetOneClassroomQuery } from "@/shared/redux/rtk-apis/classrooms/classrooms.api";
 
@@ -49,7 +48,7 @@ export default function ClassroomContainer() {
           </Tabs.List>
 
           <Tabs.Panel value="stream">
-            <StreamContainer />
+            <p> stream container here </p>
           </Tabs.Panel>
           <Tabs.Panel value="classwork">classwork tab content</Tabs.Panel>
           <Tabs.Panel value="people">
