+++ modules/classroom/containers/peopleContainer/PeopleContainer.tsx
@@ -1,13 +1,12 @@
 import { Divider } from "@mantine/core";
 import { useSelector } from "react-redux";
 
-import AddStudentOnClassroom from "@/modules/classroom/components/addStudentOnClassroom";
-import PeopleDetail from "@/modules/classroom/components/peopleDetail";
+import AddStudentToClassroom from "@/modules/classroom/components/people/addStudentToClassroom";
+import PeopleDetail from "@/modules/classroom/components/people/peopleDetail";
+import TeacherDetail from "@/modules/classroom/components/people/teacherDetail";
 import { EUserRole } from "@/shared/redux/rtk-apis/auth/auth.types";
 import type { IClassroomResponse } from "@/shared/redux/rtk-apis/classrooms/classrooms.types";
 import { TRootState } from "@/shared/redux/store";
-
-import TeacherDetail from "../../components/people/teacherDetail";
 
 export default function PeopleContainer({
   owner,
@@ -25,7 +24,7 @@ export default function PeopleContainer({
         <TeacherDetail id={owner} />
         <div className="flex flex-row justify-between gap-2">
           <h2 className="text-2xl font-bold">Students</h2>
-          {claim === EUserRole.TEACHER ? <AddStudentOnClassroom /> : <p>&nbsp;</p>}
+          {claim === EUserRole.TEACHER ? <AddStudentToClassroom /> : <p>&nbsp;</p>}
         </div>
         <Divider />
         {participants.map((participant) => (
