--- modules/classroom/containers/streamContainer/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./StreamContainer";
